chore: OrderStatus component 이름 변경
